Fix compare to unsigned int compiler warnings

+++ src/agg_font_freetype.cpp
@@ -445,12 +445,12 @@ void decompose_ft_bitmap_mono(const FT_Bitmap& bitmap,
         y += bitmap.rows;
         pitch = -pitch;
     }
-    for(i = 0; i < bitmap.rows; i++)
+    for(i = 0; i < (int) bitmap.rows; i++)
     {
         sl.reset_spans();
         bitset_iterator bits(buf, 0);
         int j;
-        for(j = 0; j < bitmap.width; j++)
+        for(j = 0; j < (int) bitmap.width; j++)
         {
             if(bits.bit()) sl.add_cell(x + j, cover_full);
             ++bits;
@@ -486,11 +486,11 @@ void decompose_ft_bitmap_gray8(const FT_Bitmap& bitmap,
         y += bitmap.rows;
         pitch = -pitch;
     }
-    for(i = 0; i < bitmap.rows; i++)
+    for(i = 0; i < (int) bitmap.rows; i++)
     {
         sl.reset_spans();
         const int8u* p = buf;
-        for(j = 0; j < bitmap.width; j++)
+        for(j = 0; j < (int) bitmap.width; j++)
         {
             if(*p) sl.add_cell(x + j, ras.apply_gamma(*p));
             ++p;
